fix: dont steal buttons

+++ src/components/buttons/cancel.js
@@ -9,6 +9,8 @@ module.exports = {
    * @param {ButtonInteraction} interaction
    */
   run: async (client, interaction) => {
+    if (interaction.user.id !== interaction.user.id) { return interaction.reply({ content: "You are not the owner of this button.", ephemeral: true }); }
+    
     await interaction.message.delete();
   },
 };
